Airflow/requirements.txt
# requirements.txt
apache-airflow-providers-databricks>=7.8.0
#databricks-sql-connector>=4.0.0
#databricks-sqlalchemy>=1.0.2
dbt-core
dbt-databricks
